+++ api/plan-dft-c2r-3d.c
@@ -20,10 +20,12 @@
 
 #include "api.h"
 
-X(plan) X(plan_dft_c2r_3d)(unsigned long nx, unsigned long ny, unsigned long nz,
-			   C *in, R *out, unsigned int flags)
+X(plan) X(plan_dft_c2r_3d) (int nx, int ny, int nz,
+                            C *in, R *out, int flags)
 {
-     unsigned long n[3];
-     n[0] = nx; n[1] = ny; n[2] = nz;
-     return X(plan_dft_c2r)(3, n, in, out, flags);
+     int n[3];
+     n[0] = nx;
+     n[1] = ny;
+     n[2] = nz;
+     return X(plan_dft_c2r) (3, n, in, out, flags);
 }
