+++ api/plan-dft-r2c-3d.c
@@ -20,10 +20,12 @@
 
 #include "api.h"
 
-X(plan) X(plan_dft_r2c_3d)(unsigned long nx, unsigned long ny, unsigned long nz,
-			   R *in, C *out, unsigned int flags)
+X(plan) X(plan_dft_r2c_3d) (int nx, int ny, int nz,
+                            R *in, C *out, int flags)
 {
-     unsigned long n[3];
-     n[0] = nx; n[1] = ny; n[2] = nz;
-     return X(plan_dft_r2c)(3, n, in, out, flags);
+     int n[3];
+     n[0] = nx;
+     n[1] = ny;
+     n[2] = nz;
+     return X(plan_dft_r2c) (3, n, in, out, flags);
 }
