+++ api/plan-many-r2r.c
@@ -21,28 +21,32 @@
 #include "api.h"
 #include "rdft.h"
 
-rdft_kind *X(map_r2r_kind)(uint rank, const X(r2r_kind) *kind);
+rdft_kind *X(map_r2r_kind) (int rank, const X(r2r_kind) * kind);
 
 #define N0(nembed) ((nembed) ? (nembed) : n)
 
-X(plan) X(plan_many_r2r)(unsigned int rank, const unsigned long *n,
-			 unsigned long howmany,
-			 R *in, const unsigned long *inembed,
-			 long istride, long idist,
-			 R *out, const unsigned long *onembed,
-			 long ostride, long odist,
-			 const X(r2r_kind) *kind, unsigned int flags)
+X(plan) X(plan_many_r2r) (int rank, const int *n,
+                          int howmany,
+                          R *in, const int *inembed,
+                          int istride, int idist,
+                          R *out, const int *onembed,
+                          int ostride, int odist,
+                          const X(r2r_kind) * kind, int flags)
 {
      X(plan) p;
-     rdft_kind *k = X(map_r2r_kind)(rank, kind);
-     
-     p = X(mkapiplan)(
-	  flags,
-	  X(mkproblem_rdft_d)(X(mktensor_rowmajor)(rank, n, 
-						   N0(inembed), N0(onembed),
-						   istride, ostride),
-			      X(mktensor_1d)(howmany, idist, odist), 
-			      in, out, k));
-     X(ifree0)(k);
+     rdft_kind *k = X(map_r2r_kind) (rank, kind);
+
+     p = X(mkapiplan) (flags,
+                       X(mkproblem_rdft_d) (X(mktensor_rowmajor) (rank, n,
+                                            N0
+                                            (inembed),
+                                            N0
+                                            (onembed),
+                                            istride,
+                                            ostride),
+                                            X(mktensor_1d) (howmany, idist,
+                                                            odist), in,
+                                            out, k));
+     X(ifree0) (k);
      return p;
 }
